requirements.txt
anthropic==0.49.0
eventlet==0.39.1
Flask==3.1.0
Flask_SocketIO==5.5.1
openai==1.70.0
python-dotenv==1.1.0
setuptools==75.8.2
typing_extensions==4.13.1
webcolors==24.11.1
